Modificaciòn forma de validar JWT y permisos en el endpoint department

+++ src/graphql/department/department.resolvers.js
@@ -1,41 +1,38 @@
-import { departmentService } from '../../services/department.service.js';
-import boom from '@hapi/boom';
-import checkRoles from '../../utils/auth/checkRolesGql.js';
+import DepartmentService from '../../services/department.service.js';
 import { checkJwtGql } from '../../utils/auth/checkJwtGql.js';
-
-const service = new departmentService();
+import { validateUserIsAdmin } from '../../utils/auth/userPermissions.js';
+const service = new DepartmentService();
 
 // Retrieves a department by name from the database.
 export const departmetByName = async (_, { departmentName }, context) => {
-  const user = await context.authenticate('jwt', { session: false });
-  if (!user) {
-    throw boom.unauthorized('Unauthorized');
-  }
-  checkRoles(user, 'customer');
+  await validateUserIsAdmin(context);
   return service.findByName(departmentName);
 };
 
 // Retrieves a department by ID from the database.
-export const departmentById = (_, { idDepartment }) => {
+export const departmentById = async (_, { idDepartment }, context) => {
+  await checkJwtGql(context);
   return service.findById(idDepartment);
 };
 
 // Retrieves all departments from the database.
-export const allDepartments = () => {
+export const allDepartments = async (_, __, context) => {
+  await validateUserIsAdmin(context);
   return service.find();
 };
 
 export const addDepartment = async (_, { dto }, context) => {
-  const user = await checkJwtGql(context);
-  checkRoles(user, 'admin');
+  await validateUserIsAdmin(context);
   const { departmentName } = dto;
   return service.create(departmentName);
 };
 
-export const updateDepartment = (_, args) => {
+export const updateDepartment = async (_, args, context) => {
+  await validateUserIsAdmin(context);
   return service.update(args);
 };
 
-export const deleteDepartment = (_, { idDepartment }) => {
+export const deleteDepartment = async (_, { idDepartment }, context) => {
+  await validateUserIsAdmin(context);
   return service.delete(idDepartment);
 };
